+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ### 0.7.3 (unreleased)
 
+- New features
+  - `ConfigReader` and `ConfigWriter` for `java.math.BigDecimal`, `java.math.BigINteger`, 
+    `scala.math.BigDecimal`, and `scala.math.BigInt`
+
 - Bug fixes
   - A breaking change introduced in v0.7.1 where `loadConfigFromFiles` stopped allowing missing files was reverted.